+++ bocadillo/__init__.py
@@ -4,8 +4,6 @@ from .config import SettingsError, configure, settings
 from .errors import HTTPError
 from .injection import discover_providers, provider, useprovider
 from .middleware import Middleware
-from .plugins import plugin
-from .recipes import Recipe
 from .redirection import Redirect
 from .request import ClientDisconnect, Request
 from .response import Response
@@ -14,7 +12,6 @@ from .sse import server_event
 from .staticfiles import static
 from .templates import Templates
 from .testing import LiveServer, create_client
-from .views import view
 from .websockets import WebSocket, WebSocketDisconnect
 
 __version__ = "0.16.2"